% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function co2_summary().R
\name{co2_summary}
\alias{co2_summary}
\title{Summary statistics for CO2 emissions by country}
\usage{
co2_summary(df, countries = NULL, start_year = NULL, end_year = NULL)
}
\arguments{
\item{df}{The CO2 dataset (output of load_co2_data)}

\item{countries}{Optional vector of countries to include}

\item{start_year}{Filter data from this year onward}

\item{end_year}{Filter data up to this year}
}
\value{
A data frame with mean, median, and sd of CO2 emissions by country
}
\description{
Summary statistics for CO2 emissions by country
}
