% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function load_co2_data().R
\name{load_co2_data}
\alias{load_co2_data}
\title{Load the CO2 dataset included in the package}
\usage{
load_co2_data()
}
\value{
A cleaned data frame with country, code, year, and CO2 emissions.
}
\description{
Load the CO2 dataset included in the package
}
