% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function plot_co2_trend().R
\name{plot_co2_trend}
\alias{plot_co2_trend}
\title{Plot CO2 emissions over time}
\usage{
plot_co2_trend(df, countries)
}
\arguments{
\item{df}{The CO2 dataset}

\item{countries}{Vector of countries to plot}
}
\value{
A ggplot line plot
}
\description{
Plot CO2 emissions over time
}
